feat: refactor TableSlideover component by moving types and constants to separate files for better organization
